bugfix: c-DEFUN doc-syntax error
The C macro DEFUN has a special syntax that enables
documentation to be built. This is parsed by lib-src/make-docfile.
A parse error will go silently ignored, which makes the functions
entry in globals.h missing.

This in turn will make the compiler use 32bit code due to an
unknown function signature. And that in turn enables bad-pointer
usage.