fix sh-smie--default-backward-token, use syntax-ppss
add workaround in sh-smie--default-backward-token
to use syntax-ppss (which works correctly) to detect if we're still in a comment after forward-comment fails, and if so, jumps directly to the comment start position.

when using forward-comment to skip backward over comments, but forward-comment fails on
unterminated comments at EOF (which is a limitation in the C implementation).
Then when inserts without a newline, the comment is unterminated. When forward-comment tries to scan backward from EOF, it looks for a comment ender character
(like newline), doesn't find one, and fails.